Restructure modules to minimize dependency on mesos jar
MesosClient for some time has been decoupled from the protos being sent back
and forth between mesos, this change extends that decoupling to
MesosSchedulerSimulation as well as the maven dependency graph itself.

* Rename mesos-rxjava-core -> mesos-rxjava-client
* Introduce new module mesos-rxjava-protobuf-client that merges the client and
  the mesos protobufs
* TestingProtos has been removed from mesos-rxjava-test due to it only being
  used in some recordio tests.
