+++ mesos-rxjava-client/pom.xml
@@ -17,20 +17,17 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
     <parent>
         <groupId>com.mesosphere.mesos.rx.java</groupId>
         <artifactId>mesos-rxjava</artifactId>
         <version>0.1.0-SNAPSHOT</version>
     </parent>
-    <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>mesos-rxjava-recordio</artifactId>
+    <artifactId>mesos-rxjava-client</artifactId>
 
-    <name>Mesos RxJava :: RecordIO Operator</name>
-
-    <description>
-        rx.Observable.Operator for extracting RecordIO messages from a stream of ByteBuf
-    </description>
+    <name>Mesos RxJava :: Client</name>
 
     <properties>
         <module.deploy.skip>false</module.deploy.skip>
@@ -38,22 +35,32 @@
 
     <dependencies>
         <dependency>
+            <groupId>com.mesosphere.mesos.rx.java</groupId>
+            <artifactId>mesos-rxjava-recordio</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.mesosphere.mesos.rx.java</groupId>
+            <artifactId>mesos-rxjava-util</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.mesosphere.mesos.rx.java</groupId>
+            <artifactId>mesos-rxjava-test</artifactId>
+        </dependency>
+        <dependency>
             <groupId>io.reactivex</groupId>
             <artifactId>rxjava</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.reactivex</groupId>
+            <artifactId>rxnetty</artifactId>
         </dependency>
         <dependency>
             <groupId>io.netty</groupId>
             <artifactId>netty-codec-http</artifactId>
         </dependency>
-        <!--suppress MavenDuplicateDependenciesInspection we're changing the scope to test instead of compile -->
         <dependency>
-            <groupId>org.apache.mesos</groupId>
-            <artifactId>mesos</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.mesosphere.mesos.rx.java</groupId>
-            <artifactId>mesos-rxjava-test</artifactId>
+            <groupId>io.netty</groupId>
+            <artifactId>netty-transport-native-epoll</artifactId>
         </dependency>
     </dependencies>
