+++ mesos-rxjava-client/src/main/java/com/mesosphere/mesos/rx/java/SinkSubscriber.java
