+++ mesos-rxjava-client/src/test/java/com/mesosphere/mesos/rx/java/SubscriberDecoratorTest.java
