+++ mesos-rxjava-client/src/test/resources/logback-test.xml
