--- mesos-rxjava-core/pom.xml
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~    Copyright (C) 2015 Mesosphere, Inc
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~    http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>com.mesosphere.mesos.rx.java</groupId>
-        <artifactId>mesos-rxjava</artifactId>
-        <version>0.1.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>mesos-rxjava-core</artifactId>
-
-    <name>Mesos RxJava Client :: Core</name>
-
-    <properties>
-        <module.deploy.skip>false</module.deploy.skip>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.mesosphere.mesos.rx.java</groupId>
-            <artifactId>mesos-rxjava-recordio</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.mesosphere.mesos.rx.java</groupId>
-            <artifactId>mesos-rxjava-util</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.mesosphere.mesos.rx.java</groupId>
-            <artifactId>mesos-rxjava-test</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.mesos</groupId>
-            <artifactId>mesos</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.reactivex</groupId>
-            <artifactId>rxjava</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.reactivex</groupId>
-            <artifactId>rxnetty</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.netty</groupId>
-            <artifactId>netty-codec-http</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.netty</groupId>
-            <artifactId>netty-transport-native-epoll</artifactId>
-        </dependency>
-    </dependencies>
-
-</project>
