+++ mesos-rxjava-example/mesos-rxjava-example-framework/pom.xml
@@ -26,7 +26,7 @@
 
     <artifactId>mesos-rxjava-example-framework</artifactId>
 
-    <name>Mesos RxJava Client :: Example :: Sleepy Framework</name>
+    <name>Mesos RxJava :: Example :: Sleepy Framework</name>
 
     <description>
         A simple framework that launches sleep tasks to exercise the client and the mesos http api.
@@ -40,7 +40,7 @@
         </dependency>
         <dependency>
             <groupId>com.mesosphere.mesos.rx.java</groupId>
-            <artifactId>mesos-rxjava-core</artifactId>
+            <artifactId>mesos-rxjava-protobuf-client</artifactId>
         </dependency>
         <dependency>
             <groupId>com.mesosphere.mesos.rx.java</groupId>
