+++ mesos-rxjava-protobuf-client/pom.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-  ~    Copyright (C) 2015 Mesosphere, Inc
+  ~    Copyright (C) 2016 Mesosphere, Inc
   ~
   ~ Licensed under the Apache License, Version 2.0 (the "License");
   ~ you may not use this file except in compliance with the License.
@@ -17,26 +17,31 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
     <parent>
         <groupId>com.mesosphere.mesos.rx.java</groupId>
         <artifactId>mesos-rxjava</artifactId>
         <version>0.1.0-SNAPSHOT</version>
     </parent>
-    <modelVersion>4.0.0</modelVersion>
-    <packaging>pom</packaging>
 
-    <groupId>com.mesosphere.mesos.rx.java.example</groupId>
-    <artifactId>mesos-rxjava-example</artifactId>
+    <artifactId>mesos-rxjava-protobuf-client</artifactId>
 
-    <name>Mesos RxJava :: Example</name>
+    <name>Mesos RxJava :: Protobuf Client</name>
 
-    <description>
-        A collection of examples (frameworks, deployable tests etc)
-    </description>
+    <properties>
+        <module.deploy.skip>false</module.deploy.skip>
+    </properties>
 
-    <modules>
-        <module>mesos-rxjava-example-framework</module>
-    </modules>
-
+    <dependencies>
+        <dependency>
+            <groupId>com.mesosphere.mesos.rx.java</groupId>
+            <artifactId>mesos-rxjava-client</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.mesos</groupId>
+            <artifactId>mesos</artifactId>
+        </dependency>
+    </dependencies>
 
 </project>
